Overhaul to a less safe design
Safer design was causing internal compiler errors so I gave up. This
version still has a bug in extracting the permutation from a coordinate.

+++ src/util.rs
@@ -1,7 +1,7 @@
-pub const fn factorial(x: u8) -> usize {
+pub const fn factorial(x: usize) -> usize {
     if x < 2 { return 1; }
-    let mut t = x as usize;
-    let mut x = x as usize - 1;
+    let mut t = x;
+    let mut x = x - 1;
 
     while x > 1 {
         t *= x;
